pyproject.toml
[project]
name = "wheretf-did-my-monies-go"
version = "0.1.0"
description = "Agentic money manager – ingest bank statements, normalize transactions, interactive dashboards + LLM chat"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "pydantic>=2.12.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.40.0",
    "litellm>=1.50.0",
    "httpx>=0.27.0",
    "PyPDF2>=3.0.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "pycryptodome>=3.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"
